WIP: 작업 중 내용 커밋

+++ src/main/java/dev/mvc/reply/ReplyProc.java
@@ -3,9 +3,11 @@ package dev.mvc.reply;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Service;
 
 @Service("dev.mvc.reply.ReplyProc")
+@Primary
 public class ReplyProc implements ReplyProcInter {
 
   @Autowired
@@ -38,12 +40,37 @@ public class ReplyProc implements ReplyProcInter {
   /** 댓글 삭제 */
   @Override
   public int delete(int reply_no) {
-    return replyDAO.delete(reply_no);
+
+    //  대댓글 개수 조회
+    int childCnt = replyDAO.countChildren(reply_no);
+
+    if (childCnt > 0) {
+      return replyDAO.softDelete(reply_no);
+    } else {
+      return replyDAO.delete(reply_no);
+    }
   }
 
   /** 댓글 1건 조회 (수정용) */
   @Override
   public ReplyVO read(int reply_no) {
     return replyDAO.read(reply_no);
+  }
+
+  @Override
+  public int updateHiddenYn(int replyNo, String hiddenYn) {
+    return replyDAO.updateHiddenYn(replyNo, hiddenYn);
+  }
+
+  /** 대댓글 개수 조회 */
+  @Override
+  public int countChildren(int reply_no) {
+    return replyDAO.countChildren(reply_no);
+  }
+
+  /** 소프트 삭제: deleted_yn = 'Y' */
+  @Override
+  public int softDelete(int reply_no) {
+    return replyDAO.softDelete(reply_no);
   }
 }
